alumnirekisteri/requirements.txt
Django==1.9
Markdown==2.6.7
PyMySQL==0.7.9
chardet==2.2.1
colorama==0.2.5
command-not-found==0.3
django-audit-log==0.7.0
django-filter==1.0.1
django-pipeline==1.3.23
djangorestframework==3.5.3
futures==2.1.6
html5lib==0.999
language-selector==0.1
mysqlclient==1.3.9
pycurl==7.19.3
pygobject==3.12.0
requests==2.2.1
six==1.5.2
ufw==0.34-rc-0ubuntu2
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.7.1
wheel==0.24.0
setuptools
